requirements.txt
astroid==2.2.5
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
console-menu==0.6.0
get==2019.4.13
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
numpy==1.17.2
pandas==0.25.1
pkg-resources==0.0.0
post==2019.4.13
praw==6.3.1
prawcore==1.0.1
public==2019.4.13
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.3
typed-ast==1.4.0
update-checker==0.16
urllib3==1.25.5
websocket-client==0.56.0
wrapt==1.11.2
